feat: Take in profile at the top level of account management
