+++ blocks/identity-block/features/account-management/_children/EmailEditableFieldContainer.jsx
@@ -2,45 +2,16 @@ import React, { useEffect, useState } from 'react';
 import { useFusionContext } from 'fusion:context';
 import getProperties from 'fusion:properties';
 import getTranslatedPhrases from 'fusion:intl';
-import { useIdentity } from '../../..';
 import EditableFieldPresentational from '../../../components/EditableFormInputField';
 import FormInputField, { FIELD_TYPES } from '../../../components/FormInputField';
 
-function EmailEditableFieldContainer() {
-  const [email, setEmail] = useState('');
-  const [error, setError] = useState('');
-
-  const { arcSite, isAdmin } = useFusionContext();
+function EmailEditableFieldContainer({ error, email }) {
+  const { arcSite } = useFusionContext();
   const { locale } = getProperties(arcSite);
   const phrases = getTranslatedPhrases(locale);
 
   const formEmailLabel = phrases.t('identity-block.email');
   const emailRequirements = phrases.t('identity-block.email-requirements');
-
-  const { Identity } = useIdentity();
-
-  useEffect(() => {
-    const getProfile = () => Identity
-      .getUserProfile()
-      .then((profileObject) => {
-        const { email: loggedInEmail } = profileObject;
-
-        if (loggedInEmail) {
-          setEmail(loggedInEmail);
-        } else {
-          setError('No email found');
-        }
-      })
-      .catch((e) => setError(e.message));
-
-    if (!isAdmin) {
-      Identity.isLoggedIn().then((isLoggedIn) => {
-        if (isLoggedIn) {
-          getProfile();
-        }
-      });
-    }
-  }, [setEmail, Identity, isAdmin]);
 
   return (
     <EditableFieldPresentational
